✨ Live refresh tickets
Currently it's polling. Could be improved by having the server be aware of what tickets are currently being displayed for the user, and then doing it on a per-ticket/row basis as opposed to the entire tbody. Could also utilise websockets or SSE to reduce the amount of requests that have to be made. However this seems to work fairly well for now.