🏗️ Don't use struct for create/update tickets
Prevents the expectation that we will update the repository version
with whatever is in the struct given. i.e. we shouldn't bypass Reveal()
or Vote() just by changing the struct and passing it in.

Additionally, since NewTicket() takes an ID parameter, we now ensure
Ticket's aren't created without an ID by accident.

It also makes the Create/Update interface a bit nicer for consumers.
Expecting consumers to manage a struct, and to not mutate
incorrect/irrelevant fields was a bit clunky. I suppose if we need to
update more fields/parameters in the future we could provide an Options
struct instead of adding more parameters.